+++ src/x12sdk/__init__.py
@@ -9,4 +9,4 @@ from dotenv import load_dotenv
 
 load_dotenv()
 
-__version__ = "1.2.0"
+__version__ = "2.0.0"